fix x86_64 build.

+++ Source/BubbleHelp/BubbleHelper.cpp
@@ -33,7 +33,7 @@
 
 #include "BubbleHelper.h"
 
-long BubbleHelper::runcount = 0;
+int32 BubbleHelper::runcount = 0;
 
 /*============================================================================================================*/
 
@@ -159,7 +159,7 @@ char *BubbleHelper::GetHelp (BView *view)
 
 /*============================================================================================================*/
 
-long BubbleHelper::_helper (void *arg)
+int32 BubbleHelper::_helper (void *arg)
 {
 	((BubbleHelper*)arg)->Helper();
 	return 0;
@@ -205,7 +205,7 @@ void BubbleHelper::Helper()
 	while  (be_app_messenger.IsValid())
 	{
 		BPoint where;
-		ulong buttons;
+		uint32 buttons;
 		if (enabled)
 		{
 			if (textwin->Lock())
